# A faster BatchSelectFunctor for tf.where on CPU.