Op 'tf.where(c, t, e)' supports that 't' and 'e' are N-D tensors while 'c' is a 1D tensor, which would call BatchSelectFunctor to get the result. But its basic implementation broadcasts 'c' to the same dimension with 't' and 'e', which would get bad efficiency on CPU for large tensors.

Here a loop-based implementation would be adopted to make this operation faster on CPU. There would be no broadcast during the calculation, thus would effectively improve the performance (seeing the benchmark test 'BatchSelect').

Benchmark result below was done on a 2 x E5-2682v4 (16c @ 2.5GHz) platform:
compiling args: -mfma -mavx2 -mavx -msse4.2 -msse4.1 -O3
(old version):
Benchmark: m_1000_n_10_use_gpu_False 	 wall_time: 0.000196 s 	 Throughput: 0.409 GB/s
Benchmark: m_1000_n_100_use_gpu_False 	 wall_time: 0.000323 s 	 Throughput: 2.48 GB/s
Benchmark: m_1000_n_1000_use_gpu_False 	 wall_time: 0.000623 s 	 Throughput: 12.8 GB/s
Benchmark: m_10000_n_10_use_gpu_False 	 wall_time: 0.00031 s 	 Throughput: 2.58 GB/s
Benchmark: m_10000_n_100_use_gpu_False 	 wall_time: 0.00061 s 	 Throughput: 13.1 GB/s
Benchmark: m_10000_n_1000_use_gpu_False 	 wall_time: 0.00364 s 	 Throughput: 22 GB/s
Benchmark: m_100000_n_10_use_gpu_False 	 wall_time: 0.000602 s 	 Throughput: 13.3 GB/s
Benchmark: m_100000_n_100_use_gpu_False 	 wall_time: 0.00367 s 	 Throughput: 21.8 GB/s
Benchmark: m_100000_n_1000_use_gpu_False 	 wall_time: 0.0309 s 	 Throughput: 25.9 GB/s

(new version):
Benchmark: m_1000_n_10_use_gpu_False 	 wall_time: 0.000153 s 	 Throughput: 0.524 GB/s
Benchmark: m_1000_n_100_use_gpu_False 	 wall_time: 0.000178 s 	 Throughput: 4.49 GB/s
Benchmark: m_1000_n_1000_use_gpu_False 	 wall_time: 0.000327 s 	 Throughput: 24.5 GB/s
Benchmark: m_10000_n_10_use_gpu_False 	 wall_time: 0.000203 s 	 Throughput: 3.94 GB/s
Benchmark: m_10000_n_100_use_gpu_False 	 wall_time: 0.000344 s 	 Throughput: 23.3 GB/s
Benchmark: m_10000_n_1000_use_gpu_False 	 wall_time: 0.00123 s 	 Throughput: 64.9 GB/s
Benchmark: m_100000_n_10_use_gpu_False 	 wall_time: 0.000326 s 	 Throughput: 24.6 GB/s
Benchmark: m_100000_n_100_use_gpu_False 	 wall_time: 0.0016 s 	 Throughput: 50 GB/s
Benchmark: m_100000_n_1000_use_gpu_False 	 wall_time: 0.0166 s 	 Throughput: 48.2 GB/s