@lowintelligence 2018-07-26
I believe I have signed the CLA.